pyproject.toml
[project]
name = "rag-proj"
version = "0.1.0"
description = "RAG (Retrieval-Augmented Generation) question-answering system with Streamlit and Google Gemini"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.15",
    "google-genai>=1.26.0",
    "gradio>=5.38.2",
    "ipykernel>=6.30.0",
    "ipywidgets>=8.1.7",
    "langchain>=0.3.26",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.8",
    "langchain-huggingface>=0.3.1",
    "pypdf>=5.8.0",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.0.0",
    "streamlit>=1.47.1",
    "tiktoken>=0.9.0",
    "transformers>=4.53.3",
    "docx2txt>=0.8",
]
